+++ lib/input/http_client.go
@@ -269,9 +269,6 @@ func (h *HTTPClient) loop() {
 		mReqParseErr = h.stats.GetCounter("request.parse.error")
 		mReqSucc     = h.stats.GetCounter("request.success")
 		mCount       = h.stats.GetCounter("count")
-		mPartsCount  = h.stats.GetCounter("parts.count")
-		mSendErr     = h.stats.GetCounter("send.error")
-		mSendSucc    = h.stats.GetCounter("send.success")
 	)
 
 	defer func() {
@@ -313,7 +310,6 @@ func (h *HTTPClient) loop() {
 
 			if msgOut != nil {
 				mCount.Incr(1)
-				mPartsCount.Incr(int64(msgOut.Len()))
 				mRcvd.Incr(1)
 				mPartsRcvd.Incr(int64(msgOut.Len()))
 			}
@@ -331,12 +327,9 @@ func (h *HTTPClient) loop() {
 				if !open {
 					return
 				}
-				if res.Error() != nil {
-					mSendErr.Incr(1)
-				} else {
+				if res.Error() == nil {
 					tracing.FinishSpans(msgOut)
 					msgOut = nil
-					mSendSucc.Incr(1)
 				}
 			case <-h.closeChan:
 				return