+++ lib/output/writer/http_client.go
@@ -80,8 +80,7 @@ func NewHTTPClient(
 		client.OptSetCloseChan(h.closeChan),
 		client.OptSetLogger(h.log),
 		client.OptSetManager(mgr),
-		// TODO: V3 change the metric path to 'client'
-		client.OptSetStats(metrics.Namespaced(h.stats, "output.http_client")),
+		client.OptSetStats(metrics.Namespaced(h.stats, "client")),
 	); err != nil {
 		return nil, err
 	}